Remove delete of publish window, which causes a crash

+++ DMHelper/src/audiotracksyrinscape.h
@@ -17,7 +17,7 @@ public slots:
     virtual void play() override;
     virtual void stop() override;
     virtual void setMute(bool mute) override;
-    virtual void setVolume(int volume) override;
+    virtual void setVolume(float volume) override;
     virtual void setRepeat(bool repeat) override;
 
 signals:
