+++ DMHelper/src/audiotrackyoutube.h
@@ -21,13 +21,13 @@ public:
     virtual bool isPlaying() const override;
     virtual bool isRepeat() const override;
     virtual bool isMuted() const override;
-    virtual int getVolume() const override;
+    virtual float getVolume() const override;
 
 public slots:
     virtual void play() override;
     virtual void stop() override;
     virtual void setMute(bool mute) override;
-    virtual void setVolume(int volume) override;
+    virtual void setVolume(float volume) override;
     virtual void setRepeat(bool repeat) override;
 
 signals:
